+++ src/Common/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Common
+{
+    public class Class1
+    {
+    }
+}
