+++ src/EventHubs/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace EventHubs
+{
+    public class Class1
+    {
+    }
+}
