+++ src/NotificationHub/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.DependencyValidation.Analyzers" version="0.10.0" targetFramework="net461" />
+</packages>
