+++ src/Relay/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Relay
+{
+    public class Class1
+    {
+    }
+}
