+++ src/ServiceBus/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace ServiceBus
+{
+    public class Class1
+    {
+    }
+}
