+++ src/ServiceBusExplorer.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2027
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29001.49
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceBusExplorer", "ServiceBusExplorer\ServiceBusExplorer.csproj", "{32754F39-E353-4607-94FE-B9B9ACFD58EF}"
 EndProject
@@ -15,6 +15,22 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceBusExplorer.Tests", "ServiceBusExplorer.Tests\ServiceBusExplorer.Tests.csproj", "{41D3C27D-37FF-43B8-AAC0-4F483E507698}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceBus", "ServiceBus\ServiceBus.csproj", "{6E3C130A-6B47-47D1-84DF-7270977E8857}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Relay", "Relay\Relay.csproj", "{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventHubs", "EventHubs\EventHubs.csproj", "{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NotificationHub", "NotificationHub\NotificationHub.csproj", "{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ToMigrate", "ToMigrate\ToMigrate.csproj", "{82070314-A7E4-437E-81DE-BDB558D2AF15}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Utilities", "Utilities\Utilities.csproj", "{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}"
+EndProject
+Project("{F088123C-0E9E-452A-89E6-6BA2F21D5CAC}") = "DependencyValidation1", "DependencyValidation1\DependencyValidation1.modelproj", "{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,6 +56,70 @@ Global
 		{41D3C27D-37FF-43B8-AAC0-4F483E507698}.Release|Any CPU.Build.0 = Release|Any CPU
 		{41D3C27D-37FF-43B8-AAC0-4F483E507698}.Release|x86.ActiveCfg = Release|Any CPU
 		{41D3C27D-37FF-43B8-AAC0-4F483E507698}.Release|x86.Build.0 = Release|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Debug|x86.Build.0 = Debug|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Release|x86.ActiveCfg = Release|Any CPU
+		{6E3C130A-6B47-47D1-84DF-7270977E8857}.Release|x86.Build.0 = Release|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Debug|x86.Build.0 = Debug|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Release|x86.ActiveCfg = Release|Any CPU
+		{C6F6FA8E-2003-421D-8F06-AAF6BF205F01}.Release|x86.Build.0 = Release|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Debug|x86.Build.0 = Debug|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Release|x86.ActiveCfg = Release|Any CPU
+		{7EB5EA53-5031-4E4E-B78C-38D370AF9A9F}.Release|x86.Build.0 = Release|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Debug|x86.Build.0 = Debug|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Release|x86.ActiveCfg = Release|Any CPU
+		{A82872B5-5A0E-49ED-BD6A-DAAC7B07A2FF}.Release|x86.Build.0 = Release|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Debug|x86.Build.0 = Debug|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Release|Any CPU.Build.0 = Release|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Release|x86.ActiveCfg = Release|Any CPU
+		{82070314-A7E4-437E-81DE-BDB558D2AF15}.Release|x86.Build.0 = Release|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Debug|x86.Build.0 = Debug|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Release|x86.ActiveCfg = Release|Any CPU
+		{93E8BC02-BCD2-4867-A990-FEC96D24E9EE}.Release|x86.Build.0 = Release|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Debug|x86.Build.0 = Debug|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Release|x86.ActiveCfg = Release|Any CPU
+		{D1AE04A9-EE81-4A2C-B235-DF9AAD821012}.Release|x86.Build.0 = Release|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Debug|x86.Build.0 = Debug|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Release|x86.ActiveCfg = Release|Any CPU
+		{61DB02D0-CBC4-425A-9714-BAAEC06AA3F3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
