+++ src/ServiceBusExplorer/Controls/HandleNotificationHubControl.cs
@@ -39,6 +39,7 @@ using Microsoft.Azure.NotificationHubs;
 using Microsoft.Azure.NotificationHubs.Messaging;
 using Microsoft.Azure.ServiceBusExplorer.Forms;
 using Microsoft.Azure.ServiceBusExplorer.Helpers;
+using Microsoft.Azure.ServiceBusExplorer.UIHelpers;
 using Microsoft.Azure.ServiceBusExplorer.Properties;
 
 #endregion
