+++ src/ServiceBusExplorer/Forms/RegistrationForm.cs
@@ -30,6 +30,7 @@ using System.Reflection;
 using System.Windows.Forms;
 using Microsoft.Azure.NotificationHubs;
 using Microsoft.Azure.ServiceBusExplorer.Helpers;
+using Microsoft.Azure.ServiceBusExplorer.UIHelpers;
 
 #endregion
