+++ src/ServiceBusExplorer/Helpers/ServiceBusHelper.cs
@@ -40,6 +40,7 @@ using Microsoft.Azure.NotificationHubs;
 using Microsoft.Azure.ServiceBusExplorer.Controls;
 using Microsoft.Azure.ServiceBusExplorer.Forms;
 using Microsoft.Azure.ServiceBusExplorer.Helpers;
+using Microsoft.Azure.ServiceBusExplorer.UIHelpers;
 using Microsoft.Azure.ServiceBusExplorer.Enums;
 #endregion
