+++ src/ServiceBusExplorer/UIHelpers/CustomCollectionEditor.cs
@@ -23,18 +23,17 @@
 
 using System;
 using System.ComponentModel;
-using System.Drawing;
 using System.Drawing.Design;
-using System.Linq;
 using System.Windows.Forms;
 using System.Windows.Forms.Design;
 using Microsoft.Azure.ServiceBusExplorer.Forms;
+using Microsoft.Azure.ServiceBusExplorer.Helpers;
 
 #endregion
 
-namespace Microsoft.Azure.ServiceBusExplorer.Helpers
+namespace Microsoft.Azure.ServiceBusExplorer.UIHelpers
 {
-    public class CustomTextEditor : UITypeEditor
+    public class CustomCollectionEditor : UITypeEditor
     {
         #region Public Methods
         public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
@@ -45,31 +44,29 @@ namespace Microsoft.Azure.ServiceBusExplorer.Helpers
         public override object EditValue(ITypeDescriptorContext context, IServiceProvider provider, object value)
         {
             var service = provider.GetService(typeof(IWindowsFormsEditorService)) as IWindowsFormsEditorService;
-            if (service != null && context.Instance is CustomObject)
+            if (service != null && value != null)
             {
-                var customObject = context.Instance as CustomObject;
                 
-                var valueAsString = value as string;
                 // ReSharper disable SuspiciousTypeConversion.Global
                 var gridItem = provider as GridItem;
                 // ReSharper restore SuspiciousTypeConversion.Global
-                
-
-                var propertyName = "Value";
+                var propertyName = value.GetType().Name;
                 if (gridItem != null)
                 {
                     propertyName = gridItem.Label;
-                    if (customObject.Properties.All(p => p.Name != gridItem.Label && p.Type != typeof(string)))
-                    {
-                        return string.Empty;
-                    }
                 }
-                using (var form = new TextForm(string.Format("Edit {0}", propertyName), propertyName, valueAsString, true))
+                var label = propertyName;
+                var oldJson = JsonSerializerHelper.Serialize(value);
+                using (var form = new CollectionEditorForm(string.Format("Edit {0}", propertyName), label, value))
                 {
-                    form.Size = new Size(600, 320);
                     if (service.ShowDialog(form) == DialogResult.OK)
                     {
-                        value = form.Content;
+                        if (string.Compare(oldJson,
+                                           JsonSerializerHelper.Serialize(value),
+                                           StringComparison.InvariantCulture) != 0)
+                        {
+                            value = GenericCopier<object>.DeepCopy(form.Value);
+                        }
                     }
                 }
             }
@@ -78,5 +75,7 @@ namespace Microsoft.Azure.ServiceBusExplorer.Helpers
             // ReSharper restore AssignNullToNotNullAttribute
         } 
         #endregion
+
+        
     }
 }
