+++ src/Utilities/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Utilities
+{
+    public class Class1
+    {
+    }
+}
